[CAS-1364] Making GiphyInfo internal
