--- stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/model/GiphyInfo.kt
@@ -1,9 +1,0 @@
-package com.getstream.sdk.chat.model
-
-public data class GiphyInfo(
-    val url: String,
-    val width: Int,
-    val height: Int,
-    val size: Int,
-    val frames: Int,
-)
